Adding Collection as sub topic

+++ api/views/problem.py
@@ -74,24 +74,29 @@ def one_problem(request,problem_id: int):
         problem.description = request.data.get("description",problem.description)
         problem.solution = request.data.get("solution",problem.solution)
         problem.time_limit = request.data.get("time_limit",problem.time_limit)  
+        problem.is_private = request.data.get("is_private",problem.is_private)
 
-        checked = checker(1,problem.solution,request.data['testcases'],request.data.get('time_limit',1.5))
-        if checked['has_error'] or checked['has_timeout']:
-            return Response({'detail': 'Error during editing. Your code may has an error/timeout!'},status=status.HTTP_406_NOT_ACCEPTABLE)
+        if 'testcases' in request.data:
+            checked = checker(1,problem.solution,request.data['testcases'],request.data.get('time_limit',1.5))
+            if checked['has_error'] or checked['has_timeout']:
+                return Response({'detail': 'Error during editing. Your code may has an error/timeout!'},status=status.HTTP_406_NOT_ACCEPTABLE)
 
-        testcases.delete()
-        testcase_result = []
-        for unit in checked['result']:
-            testcase = Testcase(
-                problem_id = problem,
-                input = unit['input'],
-                output = unit['output']
-            )
-            testcase.save()
-            testcase_result.append(model_to_dict(testcase))
+            testcases.delete()
+            testcase_result = []
+            for unit in checked['result']:
+                testcase = Testcase(
+                    problem_id = problem,
+                    input = unit['input'],
+                    output = unit['output']
+                )
+                testcase.save()
+                testcase_result.append(model_to_dict(testcase))
+            problem.save()
+
+            return Response({'detail': 'Problem has been edited!','problem': model_to_dict(problem),'testcase': testcase_result},status=status.HTTP_201_CREATED)
+
         problem.save()
-
-        return Response({'detail': 'Problem has been edited!','problem': model_to_dict(problem),'testcase': testcase_result},status=status.HTTP_201_CREATED)
+        return Response({'detail': 'Problem has been edited!','problem': model_to_dict(problem)},status=status.HTTP_201_CREATED)
 
     elif request.method == DELETE:
         problem.delete()
